Create mapping for jura forum QAs and base class for QAs
